Switch timeT from int to long.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1342 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/Event.h
@@ -41,7 +41,7 @@
 namespace Rosegarden 
 {
 
-typedef int timeT;
+typedef long timeT;
 
 /**
  * The Event class represents an event with some basic attributes and
